Use python's object model to create theme patches
This commit changes the theme patch creation logic by using pythons
dict/list object model. The final list is converted into a JSON file
via `sublime.encode_value()`.

The resulting theme patch is:
1. minified
2. uncommented

As each patch is auto-generated the user-readability is less important.
Avoiding malformed JSON by not handling any JSON format detail manually
is more important for a safe operation.

The theme patches are stored without rules key. The error message in
the console which lead to use it was caused by ST reading the file 
before it was completely written, which resulted in a json parse error
due to corrupted content. This issue is tried to workaround by 
encoding the json as text in RAM and then write it to disk.